feat(dropdown): adiciona validação e clique fora em InputDropdown
